Phase 3 progress: Fix assembly syntax, casting issues, and add missing system headers

+++ usr/src/kernel/kern/sig.c
@@ -679,9 +679,9 @@ issig(struct proc *p)
 		 * Return the signal's number, or fall through
 		 * to clear it from the pending mask.
 		 */
-		switch ((int)p->p_sigacts->ps_sigact[sig]) {
+		switch ((long)p->p_sigacts->ps_sigact[sig]) {
 
-		case SIG_DFL:
+		case (long)SIG_DFL:
 			/*
 			 * Don't take default actions on system processes.
 			 */
@@ -712,7 +712,7 @@ issig(struct proc *p)
 				return (sig);
 			/*NOTREACHED*/
 
-		case SIG_IGN:
+		case (long)SIG_IGN:
 #ifdef DIAGNOSTIC
 			/*
 			 * Masking above should prevent us ever trying
